Move Utils into Linters Package
This should fix the installation bug
